requirements.txt
cachetools==3.1.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Flask==2.0.2
google-api-python-client==1.7.11
google-auth==1.7.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.6
httplib2==0.14.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
more-itertools==7.2.0
oauth2client==4.1.3
oauthlib==3.1.1
packaging==19.2
pluggy==0.13.0
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pyparsing==2.4.5
python-dateutil==2.8.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.0
scramp==1.1.0
six==1.13.0
uritemplate==3.0.0
urllib3==1.25.6
virtualenv==16.7.7
Werkzeug==2.0.2
zipp==0.6.0
matplotlib
